pyproject.toml
[project]
name = "projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=4.2.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
